chore: fix CI pipeline, exclude reports/, resolve all clippy warnings
- Exclude reports/ from git tracking (was leaking analysis output)
- Rewrite CI workflow: use stable Rust toolchain, add clippy + tsc checks,
  add concurrency groups, npm caching, and least-privilege permissions
- Track website/package-lock.json (required by npm ci)
- Bump MSRV from 1.77 to 1.82
- Fix 93 clippy warnings: cmp_owned, &PathBuf→&Path, let-else→?,
  large error variants (Box), nested format!, identical if-blocks
- Remove stale reports/ entry from .gitattributes
- Add agent_runtime and agentic modules

+++ rust/crates/aigiscore/src/parsing/ruby.rs
@@ -198,6 +198,7 @@ fn walk_children(
     }
 }
 
+#[allow(clippy::too_many_arguments)]
 fn make_symbol(
     context: &RubyContext<'_>,
     kind: SymbolKind,
@@ -450,13 +451,13 @@ fn leaf_ruby_constant(value: &str) -> String {
 fn parameter_count(node: Node<'_>) -> usize {
     node.child_by_field_name("parameters")
         .map(|parameters| parameters.named_child_count())
-        .unwrap_or(0) as usize
+        .unwrap_or(0)
 }
 
 fn argument_count(node: Node<'_>) -> usize {
     node.child_by_field_name("arguments")
         .map(|arguments| arguments.named_child_count())
-        .unwrap_or(0) as usize
+        .unwrap_or(0)
 }
 
 #[cfg(test)]
